Add script to run e2e test locally (#11)
* Add script to run e2e test locally

* Add a more detailed comment in the test script

+++ scripts/run-test-in-region.sh
@@ -15,18 +15,23 @@ echo "Function name: ${FUNCTION_NAME}"
 
 export AWS_PROFILE=$PROFILE AWS_DEFAULT_REGION=$REGION
 
-echo "Sending request to create a function"
-cat bin/test/add-test-function.json
+if [[ -z "$SKIP_CREATE" ]]; then
+  echo "Sending request to create a function"
+  cat bin/test/add-test-function.json
 
-aws lambda create-function \
-  --cli-input-json file://bin/test/add-test-function.json \
-  --no-cli-pager ||
-    _panic "Can't create the function"
+  aws lambda create-function \
+    --cli-input-json file://bin/test/add-test-function.json \
+    --no-cli-pager ||
+      _panic "Can't create the function"
+fi
 
 response_file=$(mktemp)
 
-for i in {1..5}; do
+times=${1:-5}
+for ((i=1; i<=times; i++)); do
   echo "Calling the function ${i} time"
+
+  : > ${response_file}
 
   aws lambda invoke \
     --function-name "${FUNCTION_NAME}" \
@@ -34,14 +39,18 @@ for i in {1..5}; do
     --no-cli-pager \
     ${response_file} ||
       _panic "Can't invoke the function: $(cat ${response_file})"
+
+  [[ -z "$INVOKE_DELAY" ]] || (sleep "$INVOKE_DELAY" & wait $!)
 done
 
-echo "Sending request to delete the function"
-cat bin/test/delete-test-function.json
+if [[ -z "$SKIP_CREATE" ]]; then
+  echo "Sending request to delete the function"
+  cat bin/test/delete-test-function.json
 
-aws lambda delete-function \
-  --cli-input-json file://bin/test/delete-test-function.json \
-  --no-cli-pager ||
-    _panic "Can't remove the function"
+  aws lambda delete-function \
+    --cli-input-json file://bin/test/delete-test-function.json \
+    --no-cli-pager ||
+      _panic "Can't remove the function"
+fi
 
 echo "Done."
